requirements.txt
Flask>=3.0.0
Flask-CORS>=6.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
pyjwt>=2.8.0
cryptography>=41.0.0
google-generativeai>=0.8.0
requests>=2.31.0
